requirements.txt
discord.py>=2.0.0
python-dotenv
flask>=2.0.0
